Break apart bigger classes and extract methods
