+++ src/sqlfluff/utils/reflow/rebreak/__init__.py
@@ -1,0 +1,5 @@
+"""Methods to support ReflowSequence.rebreak().
+
+This module should contain only private methods to the
+reflow utils, or utility classes for internal use.
+"""
