+++ src/sqlfluff/utils/reflow/reindent/__init__.py
@@ -1,0 +1,1 @@
+"""Methods to support reindent."""
