+++ src/sqlfluff/utils/reflow/respace/__init__.py
@@ -1,0 +1,5 @@
+"""Methods to support ReflowSequence.respace().
+
+This module should contain only private methods to the
+reflow utils, or utility classes for internal use.
+"""
